+++ esp-hal/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- ESP32-S3: Add LCD_CAM Camera driver (#0000)
+
 ### Fixed
 
 ### Changed